Rename class to MentorJoinedChannel
Add required method in operationcode_bot.rb
Move conditional logic to check what channel was joined, and call the  appropriate MentorJoinedChannel method from there
Update tests & template to match new class name
