+++ views/event/mentor_joined_channel/welcome_message.txt.erb
